feat: add jwt auth data
